Implemented JWT Validation

+++ types.go
@@ -6,10 +6,10 @@ import (
 )
 
 type Account struct {
-	ID        int64     `json:"id"`
+	ID        int       `json:"id"`
 	FirstName string    `json:"first_name"`
 	LastName  string    `json:"last_name"`
-	Number    int64     `json:"number"`
+	Number    int       `json:"number"`
 	Balance   float64   `json:"balance"`
 	CreatedAt time.Time `json:"created_at"`
 }
@@ -20,15 +20,15 @@ type CreateAccountRequest struct {
 }
 
 type TransferRequest struct {
-	ToAccount int64 `json:"to_account"`
-	Amount    int64 `json:"amount"`
+	ToAccount int `json:"to_account"`
+	Amount    int `json:"amount"`
 }
 
 func NewAccount(FirstName, LastName string) *Account {
 	return &Account{
 		FirstName: FirstName,
 		LastName:  LastName,
-		Number:    int64(rand.Intn(1000000)),
+		Number:    int(rand.Intn(1000000)),
 		CreatedAt: time.Now().UTC(),
 	}
 }
